+++ rust/lance-encoding/src/encodings/physical/bitmap.rs
@@ -33,7 +33,7 @@ impl DenseBitmapScheduler {
 impl PageScheduler for DenseBitmapScheduler {
     fn schedule_ranges(
         &self,
-        ranges: &[Range<u32>],
+        ranges: &[Range<u64>],
         scheduler: &Arc<dyn EncodingsIo>,
         top_level_row: u64,
     ) -> BoxFuture<'static, Result<Box<dyn PrimitivePageDecoder>>> {
@@ -84,8 +84,8 @@ impl PageScheduler for DenseBitmapScheduler {
 
 struct BitmapData {
     data: Bytes,
-    bit_offset: u32,
-    length: u32,
+    bit_offset: u64,
+    length: u64,
 }
 
 struct BitmapDecoder {
@@ -95,8 +95,8 @@ struct BitmapDecoder {
 impl PrimitivePageDecoder for BitmapDecoder {
     fn decode(
         &self,
-        rows_to_skip: u32,
-        num_rows: u32,
+        rows_to_skip: u64,
+        num_rows: u64,
         _all_null: &mut bool,
     ) -> Result<Vec<BytesMut>> {
         let num_bytes = arrow_buffer::bit_util::ceil(num_rows as usize, 8);
