+++ rust/lance-encoding/src/encodings/physical/value.rs
@@ -82,7 +82,7 @@ impl ValuePageScheduler {
 impl PageScheduler for ValuePageScheduler {
     fn schedule_ranges(
         &self,
-        ranges: &[std::ops::Range<u32>],
+        ranges: &[std::ops::Range<u64>],
         scheduler: &Arc<dyn EncodingsIo>,
         top_level_row: u64,
     ) -> BoxFuture<'static, Result<Box<dyn PrimitivePageDecoder>>> {
@@ -206,8 +206,8 @@ impl ValuePageDecoder {
 impl PrimitivePageDecoder for ValuePageDecoder {
     fn decode(
         &self,
-        rows_to_skip: u32,
-        num_rows: u32,
+        rows_to_skip: u64,
+        num_rows: u64,
         _all_null: &mut bool,
     ) -> Result<Vec<BytesMut>> {
         let num_bytes = self.bytes_per_value * num_rows as u64;
